% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPeriods.R
\name{addPeriods}
\alias{addPeriods}
\title{Add periods for ICD codes}
\usage{
addPeriods(x)
}
\arguments{
\item{x}{A character vector that represents a series of ICD codes.}
}
\value{
A character vector of the same length as \code{x}. Each ICD code in 
the vector has a period added after the third digit.
}
\description{
This function add a period after the third digit of an ICD 
code and return the result as a string.
}
\examples{
bpEligICD9 <- c(7213,72190,72210,72252,7226,72293,72402,7242:7246,72470,72471, 
                72479,7385,7393,7394,8460:8469)
bpElibICD10 <- unique(getICD10(bpEligICD9))
addPeriods(bpElibICD10)
}
