% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getICD10.R
\name{getICD10}
\alias{getICD10}
\title{Convert ICD9 codes to ICD10 codes}
\usage{
getICD10(ICD9vec, return.value = c("ICD10only", "All"))
}
\arguments{
\item{ICD9vec}{A vector of ICD9 codes to convert}

\item{return.value}{By default only ICD10 codes are returned, set to 'All' 
to also return flags}
}
\value{
If \code{return.value='ICD10only'} a vector of ICD10 codes 
matching the ICD9 codes given. Otherwise, if  \code{return.value='All'}
a matrix consisting of rows from \code{ICD9to10} matching 
the requested ICD9 codes.
}
\description{
Convert ICD9 codes to ICD10 codes
}
\examples{
ICD9vec <- as.character(c(30781,33900:33999,3460:3469))
ICD10vec <- getICD10(ICD9vec)
head(ICD10vec)
}
