% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getICD-api.R
\name{getICD_api}
\alias{getICD_api}
\title{Convert between ICD9 codes and ICD10 codes}
\usage{
getICD_api(code, input = c("ICD9", "ICD10"), default = TRUE)
}
\arguments{
\item{code}{A vector of ICD codes to convert.}

\item{input}{Choose ICD version, can be "ICD9" or "ICD10".}

\item{default}{Logical value. If TRUE, then the default mapping is used. Otherwise use the 
possible mappings option which would return more results.}
}
\value{
A character vector of conversion results.
}
\description{
Convert between ICD9 codes and ICD10 codes
}
\details{
The function utilizes HIPAASpace api to convert diagnosis codes between ICD9 and ICD10. Currently 
the function uses the mapcode option of the api.
}
