% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.I10.R
\name{group.I10}
\alias{group.I10}
\title{Group similar code by their first few characters}
\usage{
group.I10(codes, minGroup = 2, maxLength = 4)
}
\value{
A character vector of the grouping result.
}
\description{
Modified after the source code of \code{convert2regex}.
}
