% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICD9to10_api.R
\docType{data}
\name{ICD9to10_api}
\alias{ICD9to10_api}
\title{Mappings from ICD-9 codes to ICD-10 codes using existing ICD9to10_api.RData}
\usage{
data(ICD9to10)
}
\description{
\code{ICD9to10_api} is a list of mappings from ICD9 codes to ICD10 codes.  
maps ICD-9 codes to ICD-10 codes.
}
