% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputSQL.R
\name{outputSQL}
\alias{outputSQL}
\title{Output SQL files}
\usage{
outputSQL(tomlFile, numTemplate, denTemplate, years = 2013:2017,
  output.path = "", option = c("1", "2", "3"))
}
\arguments{
\item{tomlFile}{The file name of the TOML file to be used.}

\item{numTemplate}{The file name of the template SQL file for numerator.}

\item{denTemplate}{The file name of the template SQL file for denominator.}

\item{years}{A vector of desired years to be pulled from the database.}

\item{output.path}{The path to the directory where the output SQL files will be stored.}

\item{option}{Choose from 1 to 3 to specify what table to be used in conversion. 
Choose 1 for ICD9to10, 2 for ICD10to9, and 3 for using both and returning the union of
two sets of results. Default is 1.}
}
\description{
This function outputs several SQL files, based on information stored
in a TOML file and two template SQL files.
}
\details{
For each input year, the output files 
include one for the numerator, one for the eligible and one for the exclusions. 
For year before 2015, the eligible file and exclusion file will use ICD-9 codes. 
Whereas for year after 2015, they will use ICD-10 codes. And for year 2015, they will
use both, so there will be two eligible files and two exclusion file. 
If the toml file does not contain exclusions, then no sql files will be generated for 
exclusions. This applies to eligible and numerator as well.
}
\examples{
## suppose the working directory is MPrOVE
outputSQL('./toml/BackPain.toml', './sql/pacprPX_CD-template.sql', 
          './sql/pacprDX_CD-template.sql', output.path = './sql/pacpr/BackPain/')
}
