% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removex.R
\name{removex}
\alias{removex}
\title{Remove the x from an ICD-9 code string}
\usage{
removex(string)
}
\arguments{
\item{string}{A vector of character strings, may or may not contain characters 
'x' and '-'.}
}
\value{
A vector of character strings. Each string only contain numeric 
characters, and is either a complete ICD-9 code or the first few digits of 
an ICD-9 code that represents all codes starting with the same substring.
}
\description{
Given a character string that represents an ICD-9 code or a range
of codes, \code{removex()} removes the x in the string and returns a code or a
code range that be used directly as input of \code{getICD10()}.
}
\examples{
removex(c("7213", "7213x", "14xx-208xx"))
removex("015xx-12xxx")
}
