% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toml2code.R
\name{toml2code}
\alias{toml2code}
\title{Read in a TOML file and output a codelist for \code{writeSQL} function}
\usage{
toml2code(file, minGroup = 2, maxLength = 4, input = c("ICD9", "ICD10"),
  output = c("ICD9", "ICD10"), option = c("1", "2", "3"))
}
\arguments{
\item{file}{A file path to the TOML file.}

\item{minGroup}{The minimum number of occurrences for a starting string pattern
to be put in the \code{$LIKE} list.}

\item{maxLength}{The maximum length of starting strings considered in \code{$LIKE}.}

\item{input, output}{Specify the version of ICD codes in the input or output. Choose 
between 'ICD9' and 'ICD10'.}

\item{option}{Choose from 1 to 3 to specify what table to be used in conversion. 
Choose 1 for ICD9to10, 2 for ICD10to9, and 3 for using both and returning the union of
two sets of results. Default is 1.}
}
\value{
A list object with three elements. 
\item{num}{code list of the numerator}
\item{elig}{code list of the inclusion}
\item{excl}{code list of the exclusion}
Each element is a code list that has two elements, \code{LIKE} and \code{IN}. And 
they can be fed into function \code{writeSQL} as the \code{codeList} parameter.
}
\description{
Read in a TOML file and output a codelist for \code{writeSQL} function
}
