% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2regex.R
\name{convert2regex}
\alias{convert2regex}
\title{Create "REGEXP_LIKE" SQL commands}
\usage{
convert2regex(nm, str, colName)
}
\arguments{
\item{nm}{A character vector of starting strings of ICD10 codes.}

\item{str}{A specific starting string that is used to match the ones in \code{nm}.
Only the matching ones in \code{nm} are considered to create the output SQL 
commands.}

\item{colName}{The corresponding column name to put in "REGEXP_LIKE" commands.}
}
\value{
A character string of a series of "REGEXP_LIKE" SQL commands.
}
\description{
This function creates a series of "REGEXP_LIKE" SQL commands, given 
a set of starting strings of ICD codes and a starting pattern to match.
}
\details{
\code{str} is usually a very short string. The function first extracts 
all the strings in \code{nm} that start with \code{str}. Then those strings are
divided into groups that share the first 1, 2, or 3 letters, with only the last
letter being different. For each of these groups, a SQL command like this is 
created, "OR REGEXP_LIKE (DX_CD,'^I0[1,2,5,6,7,8]','im')".
}
\examples{
bpExclICD9 <- c(1400:2089,2300:2399,8000:8399,8500:8549,8600:8699,
                9050:9099,92611,92612,929,9520:9529,9580:9599,3040:3042,
                3044,3054:3057,34460,7292,4210,4211,4219,0380:0389,01,
                7300:7309,7832,78079,7808,2859)
bpExclICD10 <- getICD10(bpExclICD9)
codeList <- convertToLike(codes = bpExclICD10, dict = ICD9to10$ICD10,2,4)
nm <- names(codeList$LIKE)

cat(convert2regex(nm,'C',colName = 'DX_CD'))
}
\seealso{
\code{\link{writeSQL}}
}
